all benchmarks work

+++ tests/benchmarks/typedIteratorStresstest.cpp
@@ -8,7 +8,7 @@ template<typename T> data::MemChunk<T> makeChunk( size_t size, short slice )
 	data::MemChunk<T> ret( size, size );
 	ret.setValueAs( "rowVec", util::fvector3( {1, 0} ) );
 	ret.setValueAs( "columnVec", util::fvector3( {0, 1} ) );
-	ret.setValueAs( "indexOrigin", util::fvector3( {0, 0, slice} ) );
+	ret.setValueAs( "indexOrigin", util::fvector3{0, 0, (float)slice} );
 	ret.setValueAs( "voxelSize", util::fvector3( {1, 1, 1} ) );
 	ret.setValueAs( "sequenceNumber",0);
 	return ret;
